+++ common/src/builtins.rs
@@ -34,9 +34,6 @@ pub enum BuiltinName {
 
   #[allow(non_camel_case_types)]
   console,
-
-  #[allow(non_camel_case_types)]
-  summon,
 }
 
 pub const BUILTIN_NAMES: [&str; BuiltinName::COUNT] = [
@@ -58,7 +55,6 @@ pub const BUILTIN_NAMES: [&str; BuiltinName::COUNT] = [
   "SymbolIterator",
   "BigInt",
   "console",
-  "summon",
 ];
 
 pub const BUILTIN_COUNT: usize = BuiltinName::COUNT;
