+++ compiler/src/lib.rs
@@ -1,33 +1,18 @@
-mod arithmetic_merge;
 pub mod asm;
 mod assembler;
 pub mod assembly_parser;
 mod bristol_depth;
-mod bristol_op_strings;
-mod bytecode;
-mod bytecode_decoder;
-mod bytecode_stack_frame;
-mod circuit;
-mod circuit_builder;
-mod circuit_signal;
-mod target_accessor;
-mod get_span_text;
-mod circuit_vm;
-mod circuit_vm_branch;
-mod constants;
 mod compile;
-mod cs_function;
+mod constants;
 mod diagnostic;
 mod expression_compiler;
 mod function_compiler;
 mod gather_modules;
-mod generator;
-mod id_generator;
+mod get_span_text;
 mod ident;
 mod import_pattern;
 mod instruction;
 mod link_module;
-mod make_generator_frame;
 mod module_compiler;
 mod name_allocator;
 mod optimization;
@@ -37,8 +22,8 @@ mod scope;
 mod scope_analysis;
 mod src_hash;
 mod static_expression_compiler;
+mod target_accessor;
 mod tests;
-mod val_dynamic_downcast;
 mod visit_pointers;
 
 pub use assembler::assemble;
