+++ vm/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 num-bigint = "0.4"
 num-traits = "0.2"
 summon_common = { path = "../common" }
+bristol-circuit = { git = "https://github.com/voltrevo/bristol-circuit", rev = "288847c" }
 storage = { path = "../storage" }
 num-derive = "0.4.1"
 serde_json = "1.0"
