+++ vm/src/id_generator.rs
