+++ vm/src/lib.rs
@@ -1,16 +1,25 @@
+mod arithmetic_merge;
 mod array_higher_functions;
 mod array_methods;
 mod bigint_methods;
 pub mod binary_op;
+mod bristol_op_strings;
 mod builtins;
 mod bytecode;
 mod bytecode_decoder;
 mod bytecode_stack_frame;
 pub mod cat_stack_frame;
+pub mod circuit;
+pub mod circuit_builder;
+pub mod circuit_signal;
+pub mod circuit_vm;
+mod circuit_vm_branch;
 mod copy_counter;
+pub mod cs_function;
 mod first_stack_frame;
 mod generator;
 mod helpers;
+pub mod id_generator;
 mod iteration;
 pub mod jsx_element;
 mod make_generator_frame;
@@ -22,8 +31,8 @@ mod stack_frame;
 mod string_methods;
 mod todo_fn;
 pub mod unary_op;
+pub mod val_dynamic_downcast;
 mod val_storage;
-mod virtual_machine;
 pub mod vs_array;
 pub mod vs_class;
 mod vs_function;
@@ -42,6 +51,5 @@ pub use iteration::iteration_result::IterationResult;
 pub use iteration::return_this::RETURN_THIS;
 pub use jsx_element::is_jsx_element;
 pub use stack_frame::{CallResult, FrameStepOk, FrameStepResult, StackFrame, StackFrameTrait};
-pub use virtual_machine::VirtualMachine;
 pub use vs_symbol::VsSymbol;
 pub use vs_value::{LoadFunctionResult, ValTrait};
