+++ vm/src/make_generator_frame.rs
@@ -1,10 +1,10 @@
-use std::{any::Any, mem::take};
+use std::mem::take;
 
 use crate::{
   bytecode_stack_frame::BytecodeStackFrame,
   generator::Generator,
-  stack_frame::{CallResult, FrameStepOk, FrameStepResult, StackFrame, StackFrameTrait},
   vs_value::{ToDynamicVal, Val},
+  CallResult, FrameStepOk, FrameStepResult, StackFrame, StackFrameTrait,
 };
 
 #[derive(Clone)]
@@ -62,11 +62,11 @@ impl StackFrameTrait for MakeGeneratorFrame {
     Box::new(self.clone())
   }
 
-  fn as_any(&self) -> &dyn Any {
+  fn as_any(&self) -> &dyn std::any::Any {
     self
   }
 
-  fn as_any_mut(&mut self) -> &mut dyn Any {
+  fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
     self
   }
 }
